requirements.txt

langchain-openai==0.2.5
langgraph==0.2.44
langgraph-checkpoint-sqlite==2.0.1

tavily-python==0.5.0

praw
python-dotenv
transformers 
torch
bitsandbytes
accelerate
langchain
langchain_groq
python-telegram-bot
flask